requirements.txt
numpy
pandas
openai>=1.14.0
google-generativeai>=0.5.2
